pyproject.toml
[project]
name = "lambda-framework"
version = "0.3.0"
description = "A Python framework for managing environment-specific configuration in AWS Lambda functions, with built-in support for secrets management via AWS Secrets Manager and GitHub webhook handling."
readme = "README.md"
authors = [
    { name = "Doni Crosby", email = "doni.crosby@dbtlabs.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aws-secretsmanager-caching>=1.1.3",
    "fastapi>=0.128.0",
    "mangum>=0.20.0",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aioboto3>=13.0.0",
    "fastapi[standard]>=0.128.0",
    "githubkit>=0.14.2",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "redis>=7.0.0",
    "ruff>=0.14.10",
    "types-redis>=4.6.0.20241004",
    "pottery-semaphore",
]

[project.optional-dependencies]
cache = [
    "redis>=7.0.0",
]
eventbridge = [
    "aioboto3>=13.0.0",
]
github = [
    "githubkit>=0.14.2",
    "pottery>=3.0.1",
    "pottery-semaphore",
    "redis>=7.0.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP",  "B", "S", "A", "D", "C90", "SIM"]
ignore = ["E501", "D203", "D213"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["S101", "S106", "D104", "S105", "S603", "S607"]

[tool.uv.sources]
pottery-semaphore = { git = "https://github.com/donicrosby/pottery-semaphore.git" }

[tool.pytest.ini_options]
minversion = "9.0"
addopts = ["--asyncio-mode=auto", "-ra", "-q"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require Docker/containers)",
]
